Change reading technique
Drop the base64, does not work properly when base64 is a bigger than a certain size. With this change, we pass the URL of the pass instead of base64. To not change many things, you still pass it as {base64: url}. The reason behind this is because capacitor / webkit is pretty bad at downloading files that are passed a certain size and it truncates the data, but swift does not have this limitation.